Fix duplicate action issue
